@@ -1,9 +1,9 @@
 {
   "name": "@triangular/noto-emoji-icons",
-  "version": "5.0.3",
+  "version": "6.0.0",
   "peerDependencies": {
-    "@angular/common": "^13.0.0",
-    "@angular/core": "^13.0.0"
+    "@angular/common": "^14.0.0",
+    "@angular/core": "^14.0.0"
   },
   "homepage": "https://noto-emoji.akehir.com",
   "license" : "MIT",
